25.12.22

+++ Mini-Bot/lib/CodeOfRobot/MotorControl/MotorControl.h
@@ -1,10 +1,14 @@
 #include "Arduino.h"
+#include "analogWrite.h"
+#include "../RemoteControl/RemoteControl.h"
+
+
 class MotorControl
 {
     public:
         static void setup();
         static void loop();
-        static void driveLeft(int speed);
-        static void driveRight(int speed);
+        static void driveLeft(double speed);
+        static void driveRight(double speed);
 };
